updated getTodo route
