+++ src/routes/todo-list-edit.route.js
@@ -1,5 +1,5 @@
 import { updateTodo } from "../controllers/todo-list-edit.controller";
 
 export default (router) => {
-  router.patch('user/userId/todos/listId', updateTodo);
+  router.patch('user/:userId/todo/:listId', updateTodo);
 }
